+++ src/scripts/directives/tink-pagination.js
@@ -3,7 +3,7 @@
   try {
     module = angular.module('tink.interactivetable');
   } catch (e) {
-    module = angular.module('tink.interactivetable', ['tink.popover','tink.sorttable','ngLodash']);
+    module = angular.module('tink.interactivetable', ['tink.popover','tink.sorttable','tink.tooltip','tink.safeApply']);
   }
   module.directive('tinkPagination',['lodash',function(_){
   return{
@@ -14,7 +14,7 @@
       tinkItemsPerPage:'=',
       tinkItemsPerPageValues:'=',
       tinkCurrentPage:'=',
-      tinkChange:'=',
+      tinkChange:'&',
       tinkPaginationId:'@'
     },
     controllerAs:'ctrl',
@@ -42,12 +42,13 @@
         }
       };
 
+
       ctrl.perPageChange = function(){
         $rootScope.$broadcast('tink-pagination-'+$scope.tinkPaginationId,'loading');
         timeout(function(){
-          $scope.tinkChange({type:'perPage',value:$scope.tinkItemsPerPage},function(){
+          $scope.tinkChange({page:$scope.tinkCurrentPage,perPage:$scope.tinkItemsPerPage,next:function(){
             $rootScope.$broadcast('tink-pagination-'+$scope.tinkPaginationId,'ready');
-          });
+          }});
         },0);
       };
 
@@ -74,9 +75,9 @@
       function sendMessage(){
         $rootScope.$broadcast('tink-pagination-'+$scope.tinkPaginationId,'loading');
         timeout(function(){
-          $scope.tinkChange({type:'page',value:$scope.tinkCurrentPage},function(){
+          $scope.tinkChange({page:$scope.tinkCurrentPage,perPage:$scope.tinkItemsPerPage,next:function(){
             $rootScope.$broadcast('tink-pagination-'+$scope.tinkPaginationId,'ready');
-          });
+          }});
         },0);
       }
 
@@ -138,10 +139,16 @@
         
         rootScope.$on('tink-pagination-'+attrs.tinkPaginationKey,function(e,value){
 
+          var table;
+          if(element[0].tagName === 'TABLE'){
+            table = $(element[0]);
+          }else {
+            table = $(element).find('table');
+          }
           if(value === 'loading'){
-            $(element).find('table.table-interactive').addClass('is-loading');
+            table.addClass('is-loading');
           }else if(value === 'ready'){
-            $(element).find('table.table-interactive').removeClass('is-loading'); 
+            table.removeClass('is-loading'); 
           }
           
         });