+++ src/scripts/directives/tink-reorder.js
@@ -3,9 +3,9 @@
   try {
     module = angular.module('tink.interactivetable');
   } catch (e) {
-    module = angular.module('tink.interactivetable', ['tink.popover','tink.sorttable','ngLodash']);
+    module = angular.module('tink.interactivetable', ['tink.popover','tink.sorttable','tink.tooltip','tink.safeApply']);
   }
-  module.directive('tinkInteractiveTable',['$compile','$rootScope','$filter',function($compile,$rootScope,$filter){
+  module.directive('tinkInteractiveTable1',['$compile','$rootScope','$filter',function($compile,$rootScope,$filter){
   return{
     restrict:'EA',
     templateUrl:'templates/reorder.html',